requirement.txt
asgiref==3.8.1
Django==5.1.2
django-appconf==1.0.6
django-compressor==4.5.1
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-htmx==1.21.0
django-livereload-server==0.5.1
django-sass-processor==1.4.1
libsass==0.23.0
MarkupSafe==3.0.2
pillow==11.0.0
pygraphviz==1.14
rcssmin==1.1.2
rjsmin==1.2.2
sqlparse==0.5.1
tornado==6.4.1
watchdog==5.0.3
websocket-client==1.8.0
Werkzeug==3.0.6
